move RouteAnnouncer into its own file
